server/requirements.txt
appdirs==1.4.3
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
bleach==3.2.1
blinker==1.4
bs4==0.0.1
cachetools==4.0.0
certifi==2019.3.9
cffi==1.13.2
chardet==3.0.4
Click==7.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
gevent==20.9.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud-core==1.2.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.17
gunicorn==20.0.4
httplib2==0.17.1
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.3.4
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
lxml==4.6.2
MarkupSafe==1.1.1
meds==1215
mistune==0.8.4
mysql-connector-python==8.0.15
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.19.2
oauth2client==4.1.3
packaging==20.4
pandas==1.1.4
pandas-datareader==0.9.0
pandocfilters==1.4.2
passlib==1.7.2
pexpect==4.8.0
pickleshare==0.7.5
project==20
prometheus-client==0.8.0
prompt-toolkit==1.0.18
protobuf==3.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.7.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-fmrest==1.3.0
pytz==2019.3
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.21.0
rsa==4.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.14.0
sleekxmpp==1.3.3
soupsieve==1.9.4
sql==0.4.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
uritemplate==3.0.1
urllib3==1.24.1
virtualenv==20.0.20
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.1.0
